add AmazonGetItemWorker
